refactor: shared explorer url parsing logic
