+++ projects/observability/src/pages/explorer/explorer.component.ts
@@ -5,30 +5,24 @@ import {
   NavigationService,
   PreferenceService,
   QueryParamObject,
-  TimeDuration,
-  TimeDurationService
+  TimeDuration
 } from '@hypertrace/common';
 import { Filter, FilterAttribute, ToggleItem } from '@hypertrace/components';
-import { isEmpty, isNil } from 'lodash-es';
+import { isEmpty } from 'lodash-es';
 import { concat, EMPTY, Observable, Subject } from 'rxjs';
 import { map, take } from 'rxjs/operators';
-import { CartesianSeriesVisualizationType } from '../../shared/components/cartesian/chart';
 import {
   ExploreOrderBy,
   ExploreRequestState,
   ExploreSeries,
   ExploreVisualizationRequest
 } from '../../shared/components/explore-query-editor/explore-visualization-builder';
-import { SortDirection } from '../../shared/components/explore-query-editor/order-by/explore-query-order-by-editor.component';
 import { IntervalValue } from '../../shared/components/interval-select/interval-select.component';
 import { AttributeExpression } from '../../shared/graphql/model/attribute/attribute-expression';
 import { toFilterAttributeType } from '../../shared/graphql/model/metadata/attribute-metadata';
-import { MetricAggregationType } from '../../shared/graphql/model/metrics/metric-aggregation';
 import { GraphQlGroupBy } from '../../shared/graphql/model/schema/groupby/graphql-group-by';
 import { ObservabilityTraceType } from '../../shared/graphql/model/schema/observability-traces';
-import { GraphQlSortDirection } from '../../shared/graphql/model/schema/sort/graphql-sort-direction';
 import { SPAN_SCOPE } from '../../shared/graphql/model/schema/span';
-import { ExploreSpecificationBuilder } from '../../shared/graphql/request/builders/specification/explore/explore-specification-builder';
 import { MetadataService } from '../../shared/services/metadata/metadata.service';
 import {
   ExplorerDashboardBuilder,
@@ -37,6 +31,7 @@ import {
   ExplorerGeneratedDashboardContext,
   EXPLORER_DASHBOARD_BUILDER_FACTORY
 } from './explorer-dashboard-builder';
+import { ExplorerUrlParserService } from './explorer-url-parser.service';
 
 @Component({
   selector: 'ht-explorer',
@@ -157,10 +152,10 @@ export class ExplorerComponent {
   public constructor(
     private readonly metadataService: MetadataService,
     protected readonly navigationService: NavigationService,
-    private readonly timeDurationService: TimeDurationService,
     private readonly preferenceService: PreferenceService,
     @Inject(EXPLORER_DASHBOARD_BUILDER_FACTORY) explorerDashboardBuilderFactory: ExplorerDashboardBuilderFactory,
-    private readonly activatedRoute: ActivatedRoute
+    private readonly activatedRoute: ActivatedRoute,
+    private readonly explorerUrlParserService: ExplorerUrlParserService
   ) {
     this.explorerDashboardBuilder = explorerDashboardBuilderFactory.build();
     this.visualizationExpanded$ = this.preferenceService.get(ExplorerComponent.VISUALIZATION_EXPANDED_PREFERENCE, true);
@@ -271,9 +266,11 @@ export class ExplorerComponent {
   private mapToInitialState(param: ParamMap): InitialExplorerState {
     const series: ExploreSeries[] = param
       .getAll(ExplorerQueryParam.Series)
-      .flatMap((seriesString: string) => this.tryDecodeExploreSeries(seriesString));
+      .flatMap((seriesString: string) => this.explorerUrlParserService.tryDecodeExploreSeries(seriesString));
 
-    const interval: IntervalValue = this.decodeInterval(param.get(ExplorerQueryParam.Interval));
+    const interval: IntervalValue = this.explorerUrlParserService.decodeInterval(
+      param.get(ExplorerQueryParam.Interval)
+    );
 
     return {
       contextToggle: this.getOrDefaultContextItemFromQueryParam(param.get(ExplorerQueryParam.Scope) as ScopeQueryParam),
@@ -281,7 +278,9 @@ export class ExplorerComponent {
         ? {
             keyExpressions: param
               .getAll(ExplorerQueryParam.Group)
-              .flatMap(expressionString => this.tryDecodeAttributeExpression(expressionString)),
+              .flatMap(expressionString =>
+                this.explorerUrlParserService.tryDecodeAttributeExpression(expressionString)
+              ),
             includeRest: param.get(ExplorerQueryParam.OtherGroup) === 'true',
 
             limit: parseInt(param.get(ExplorerQueryParam.GroupLimit)!) || 5
@@ -291,7 +290,10 @@ export class ExplorerComponent {
       series: series,
       orderBy:
         interval === 'NONE'
-          ? this.tryDecodeExploreOrderBy(series[0], param.get(ExplorerQueryParam.Order) ?? undefined)
+          ? this.explorerUrlParserService.tryDecodeExploreOrderBy(
+              series[0],
+              param.get(ExplorerQueryParam.Order) ?? undefined
+            )
           : undefined
     };
   }
@@ -307,17 +309,6 @@ export class ExplorerComponent {
     return interval.toString();
   }
 
-  private decodeInterval(durationString: string | null): IntervalValue {
-    if (isNil(durationString)) {
-      return 'AUTO';
-    }
-    if (durationString === 'NONE') {
-      return durationString;
-    }
-
-    return this.timeDurationService.durationFromString(durationString) ?? 'AUTO';
-  }
-
   private encodeExploreSeries(series: ExploreSeries): string {
     return `${series.visualizationOptions.type}:${series.specification.aggregation}(${series.specification.name})`;
   }
@@ -326,59 +317,12 @@ export class ExplorerComponent {
     return `${orderBy.aggregation}(${orderBy.attribute.key}):${orderBy.direction}`;
   }
 
-  private tryDecodeExploreSeries(seriesString: string): [ExploreSeries] | [] {
-    const matches = seriesString.match(/(\w+):(\w+)\((\w+)\)/);
-    if (matches?.length !== 4) {
-      return [];
-    }
-
-    const visualizationType = matches[1] as CartesianSeriesVisualizationType;
-    const aggregation = matches[2] as MetricAggregationType;
-    const key = matches[3];
-
-    return [
-      {
-        specification: new ExploreSpecificationBuilder().exploreSpecificationForKey(key, aggregation),
-        visualizationOptions: {
-          type: visualizationType
-        }
-      }
-    ];
-  }
-
-  private tryDecodeExploreOrderBy(selectedSeries: ExploreSeries, orderByString?: string): ExploreOrderBy {
-    const matches = orderByString?.match(/(\w+)\((\w+)\):(\w+)/);
-
-    if (matches?.length !== 4) {
-      return {
-        aggregation: selectedSeries.specification.aggregation as MetricAggregationType,
-        direction: SortDirection.Desc,
-        attribute: {
-          key: selectedSeries.specification.name
-        }
-      };
-    }
-
-    return {
-      aggregation: matches[1] as MetricAggregationType,
-      direction: matches[3] as GraphQlSortDirection,
-      attribute: {
-        key: matches[2]
-      }
-    };
-  }
-
   private encodeAttributeExpression(attributeExpression: AttributeExpression): string {
     if (isEmpty(attributeExpression.subpath)) {
       return attributeExpression.key;
     }
 
     return `${attributeExpression.key}__${attributeExpression.subpath}`;
-  }
-  private tryDecodeAttributeExpression(expressionString: string): [AttributeExpression] | [] {
-    const [key, subpath] = expressionString.split('__');
-
-    return [{ key: key, ...(!isEmpty(subpath) ? { subpath: subpath } : {}) }];
   }
 }
 interface ContextToggleItem extends ToggleItem<ExplorerContextScope> {
