+++ projects/observability/src/public-api.ts
@@ -199,6 +199,8 @@ export * from './shared/dashboard/data/graphql/explore/explore-cartesian-data-so
 export * from './pages/explorer/explorer-dashboard-builder';
 export * from './pages/explorer/explorer.component';
 
+export * from './pages/explorer/explorer-url-parser.service';
+
 // Explore Data source
 export * from './shared/dashboard/data/graphql/table/explore/explore-table-data-source.model';
